requirements.txt
setuptools
wheel
spacy==2.3.2
spacy-streamlit
st-annotated-text
docx2txt
pdfplumber
pypdf2
en_core_web_sm@https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz